RTUUI.iOS: multipage scan, save, view
